Actually remove it from admin
